requirements.txt
flask
pymysql
sqlalchemy
pandas
python-dotenv
requests

